feat: Display a button element
In this commit, i created and display a button element, rendering the data
being passed from the parent in props.

Also, in the NumberButton.js file, i exported NumberButton component  which
returns a button component and passed a number prop into it.

The Number component was also exported from the Number.js and simultaneously
imported in the App.js file

within the app.js i looped through the numberState and created a button for each
of the number