feat: Create and render operator characters
In this commit, I created and rendered the individual operator character
present in the data.js file. I imported the OperatorButton component to
the Operator parent component and passed the array of objects containing
the operator character objects.

This made it possible for me to render the individual operator character
present in the data.js file.

Also, the NumberButton child component was imported into the parent Number
component. Within the Number component, an array of numbers imported from the data
file was passed into the NumberButton component and used to create the
individual numbers.

Lastly, within the App.js the Numbers and Operators components were imported
and rendered to display all numbers and operators